Factor out inttest logging setup into a separate package
