+++ integrationtests/chrome/chrome_suite_test.go
@@ -18,6 +18,8 @@ import (
 	"github.com/lucas-clemente/quic-go/internal/utils"
 	"github.com/lucas-clemente/quic-go/protocol"
 
+	_ "github.com/lucas-clemente/quic-go/integrationtests/tools/testlog"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/onsi/gomega/gexec"
