+++ integrationtests/integrationtests_suite_test.go
@@ -1,14 +1,11 @@
 package integrationtests
 
 import (
-	"flag"
 	"fmt"
 	"io"
 	"io/ioutil"
-	"log"
 	"net"
 	"net/http"
-	"os"
 	"path/filepath"
 	"runtime"
 
@@ -16,7 +13,8 @@ import (
 
 	"github.com/lucas-clemente/quic-go/h2quic"
 	"github.com/lucas-clemente/quic-go/internal/testdata"
-	"github.com/lucas-clemente/quic-go/internal/utils"
+
+	_ "github.com/lucas-clemente/quic-go/integrationtests/tools/testlog"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -35,9 +33,6 @@ var (
 	port       string
 	clientPath string
 	serverPath string
-
-	logFileName string // the log file set in the ginkgo flags
-	logFile     *os.File
 )
 
 func TestIntegration(t *testing.T) {
@@ -47,40 +42,18 @@ func TestIntegration(t *testing.T) {
 
 var _ = BeforeSuite(setupHTTPHandlers)
 
-// read the logfile command line flag
-// to set call ginkgo -- -logfile=log.txt
-func init() {
-	flag.StringVar(&logFileName, "logfile", "", "log file")
-}
-
 var _ = BeforeEach(func() {
-	// set custom time format for logs
-	utils.SetLogTimeFormat("15:04:05.000")
 	_, thisfile, _, ok := runtime.Caller(0)
 	if !ok {
 		Fail("Failed to get current path")
 	}
 	clientPath = filepath.Join(thisfile, fmt.Sprintf("../../../quic-clients/client-%s-debug", runtime.GOOS))
 	serverPath = filepath.Join(thisfile, fmt.Sprintf("../../../quic-clients/server-%s-debug", runtime.GOOS))
-
-	if len(logFileName) > 0 {
-		var err error
-		logFile, err = os.Create("./log.txt")
-		Expect(err).ToNot(HaveOccurred())
-		log.SetOutput(logFile)
-		utils.SetLogLevel(utils.LogLevelDebug)
-	}
 })
 
 var _ = JustBeforeEach(startQuicServer)
 
-var _ = AfterEach(func() {
-	stopQuicServer()
-
-	if len(logFileName) > 0 {
-		_ = logFile.Close()
-	}
-})
+var _ = AfterEach(stopQuicServer)
 
 func setupHTTPHandlers() {
 	defer GinkgoRecover()
